Test actually updating the wiki on Travis
This seems better than any kind of mock test that I could come up with. The
test will update https://github.com/sympy/sympy-bot/wiki/Release-Notes-Tests
when it is run on Travis. There are some caveats (see the README). The tests
require a personal access token for the sympy-bot user to be added to Travis,
and they can only be run on Travis, not locally or on builds from PRs from
forks.